requirements.txt
asgiref==3.8.1
Django==5.0.9
django-mssql-backend==2.8.1
djangorestframework==3.15.2
mssql-django==1.5
pyodbc==5.2.0
pytz==2024.2
sqlparse==0.5.2
tzdata==2024.2
